requirements.txt
Flask
Flask-Login
Flask-WTF
WTForms
PyMySQL
cryptography
python-dotenv
requests
google-generativeai
werkzeug
Pillow